refactor: ProjectCategory 분리
