+++ ProjectManager/ProjectManager/Model/ProjectList.swift
@@ -7,23 +7,6 @@
 
 import Foundation
 
-enum ProjectCategory {
-  case todo
-  case doing
-  case done
-
-  var description: String {
-    switch self {
-    case .todo:
-      return "TODO"
-    case .doing:
-      return "DOING"
-    case .done:
-      return "DONE"
-    }
-  }
-}
-
 struct ProjectList {
   let title: String
   let body: String
